requirements.txt
Flask==2.3.3
twilio==8.15.0
google-genai==0.9.3
